+++ learning/katas/go/core_transforms/map/lesson-info.yaml
@@ -18,6 +18,6 @@
 #
 
 content:
-- ParDo
+- pardo
 - ParDo OneToMany
 - ParDo Struct
