+++ learning/katas/go/core_transforms/map/pardo/pkg/task/task.go
