+++ learning/katas/go/core_transforms/map/pardo/task-remote-info.yaml
