+++ learning/katas/go/core_transforms/map/pardo/task.md
